Fix stats:
* update requirements to include matplotlib
* Update cv_get_stats to have an optional import
* move plot_and_save_hist to vis
