+++ requirements.txt
@@ -13,3 +13,4 @@ torch==2.3.1
 tqdm==4.66.4
 transformers==4.41.2
 CausalEstimate==0.2.2
+matplotlib>=3.8.5
